requirement.txt
python-telegram-bot==13.14
m3u8==3.4.0
requests==2.28.1